Resolved warnings for Xcode 5.1. removed extra methods.

+++ BasicStructure/Library/Extensions/NSDateExtras/NSDate+extras.m
@@ -55,7 +55,7 @@ static NSDateFormatter *displayFormatter;
 			text = @"Yesterday";
 			break;
 		default:
-			text = [NSString stringWithFormat:@"%d days ago", daysAgo];
+			text = [NSString stringWithFormat:@"%lu days ago", (unsigned long)daysAgo];
 	}
 	return text;
 }
